+++ src/main/java/com/sparta/WeatherWear/repository/CommentLikeRepository.java
@@ -4,5 +4,5 @@ package com.sparta.WeatherWear.repository;
 import com.sparta.WeatherWear.entity.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentLikeRepository extends JpaRepository<CommentLike, Integer> {
+public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
 }
